requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
gunicorn==22.0.0
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
python-jose==3.3.0
rest-framework-simplejwt==0.0.2
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
